I want to dip my brain in a can of paint thinner
Fixes the authentication issue. We were querying for columns that didn't exist in the User database and because Prisma is Prisma (this is not a complement) we weren't getting any errors relating to the issue. Fixed lines in 'api/dist/functions/auth.js', 'api/src/functions/auth.js' and 'api/src/lib/auth.js'. Holy Fuckles it's knuckles this one was a doozy